feat: change cpp build to use setup.py

+++ capsa/c_extension/include/utils.h
@@ -28,7 +28,7 @@ std::vector<std::vector<T>> generateChooseK(std::vector<T> &vector, size_t k) {
 
 template <typename T>
 void cartesianProductHelper(const std::vector<std::vector<T>> &groups,
-                            int index,
+                            size_t index,
                             std::vector<T> &current,
                             std::vector<std::vector<T>> &result) {
     if (index == groups.size()) {
